Move out MemcpyTimingRecorder from device_memcpy.cpp
